+++ js/data.js
@@ -47,7 +47,7 @@ const craeteComments = function() {
   };
 };
 
-const commentArray = () => new Array(gettingValue(1, 3)).fill(null).map(() => craeteComments());
+const commentArray = () => new Array(gettingValue(1, 4)).fill(null).map(() => craeteComments());
 
 const createPhotoBlock = function() {
   return {
